Image viewer fix.

+++ BackEnd/src/main/java/kr/ac/wku/cephalometricai/service/ImageService.java
@@ -126,11 +126,10 @@ public class ImageService {
         if(imageOptional.isEmpty()) return;
         Image image = imageOptional.get();
         Path path = Paths.get(privateCloudProperties.getPath()+"/"+image.getOwner().toString()).resolve(image.getSystemPath());
-        Files.delete(path);
+        if(Files.exists(path))  Files.delete(path);
         String jsonName = image.getSystemPath().substring(0, image.getSystemPath().lastIndexOf('.'));
         Path jsonPath = Paths.get(privateCloudProperties.getPath()+"/"+image.getOwner().toString()).resolve(jsonName + ".json");
-        if(Files.exists(jsonPath))
-            Files.delete(jsonPath);
+        if(Files.exists(jsonPath)) Files.delete(jsonPath);
         imageRepository.deleteById(id);
     }
 }
